experiments/rq3_4/pyproject.toml
[project]
name = "rq3-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruamel-yaml>=0.18.14",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]

[tool.ruff]
ignore = ["E402"]